+++ src/mlp.h
@@ -8,8 +8,8 @@ typedef enum {
 }fn_name;
 
 typedef struct {
-    void (*f)(double* inputs, double* outputs, int len);
-    void (*df)(double* inputs, double* outputs, int len);
+    void (*f)(float* inputs, float* outputs, int len);
+    void (*df)(float* inputs, float* outputs, int len);
     fn_name function_name;
 }function;
 
@@ -20,11 +20,11 @@ extern function softm;
 
 typedef struct {
     // y = a_i * x_i + b
-    double* weights;
+    float* weights;
     int n_weights;
-    double output;
-    double bias;
-    double delta; // Neuron error
+    float output;
+    float bias;
+    float delta; // Neuron error
 }neuron;
 
 // layer = list of neurons with activation function
@@ -32,9 +32,9 @@ typedef struct {
     int n_neurons;
     neuron* neurons;
     function* activation_function;
-    double* raw_outputs;
-    double* outputs; // Output of each neuron of the layer
-    double* derivatives;
+    float* raw_outputs;
+    float* outputs; // Output of each neuron of the layer
+    float* derivatives;
 }layer;
 
 typedef struct {
@@ -43,16 +43,16 @@ typedef struct {
 } MLP;
 
 
-double ranged_rand(double min, double max);
+float ranged_rand(float min, float max);
 void init_neuron(neuron* neuron, int n_parameters);
 layer dense(int n_neurons, int n_parameters, function *activation_function);
-void forward(MLP *m, double* inputs, int n_inputs);
-void train(MLP *m, double* raw_inputs, double* target, double lr);
+void forward(MLP *m, float* inputs, int n_inputs);
+void train(MLP *m, float* raw_inputs, float* target, float lr);
 int get_num_parameters(MLP* mlp);
 void print_model(MLP* m);
-void print_output(MLP *m, double* input, int input_len, double *expected, int expected_len);
-void print_list(double* list, int len);
+void print_output(MLP *m, float* input, int input_len, float *expected, int expected_len);
+void print_list(float* list, int len);
 void free_model(MLP* m);
-void one_hot(double* output, int input, int n_classes);
+void one_hot(float* output, int input, int n_classes);
 void save_model(MLP* m, const char* path);
 void load_model(MLP* m, const char* path);
